refactor(converters): split `converters.py` into a package
